pyproject.toml
[tool.poetry]
name = "expert_dollup"
version = "0.1.0"
description = ""
authors = ["gabriel <gabcpp@outlook.fr>"]

[tool.poetry.dependencies]
python = "^3.9.5"
fastapi = "^0.61.2"
python-dotenv = "^0.15.0"
python-json-logger = "^2.0.4"
jsonschema = "^3.2.0"
python-multipart = "^0.0.5"
mypy-extensions = "^0.4.3"
google-cloud-storage = "^1.43.0"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
ariadne = "^0.14.0"
orjson = "^3.6.5"
PyJWT = {extras = ["crypto"], version = "^2.3.0"}
dependency-injector = "^4.40.0"

# Optional
google-cloud-firestore = {version = "^2.3.4", optional = true}
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.29", optional = true}
asyncpg = {version = "^0.25.0", optional = true}
motor = {extras = ["srv"], version = "^3.0.0", optional = true}
pymongo = {extras = ["srv"], version = "^4.2.0", optional = true}


[tool.poetry.extras]
mongo = ["motor", "pymongo"]
postgre = ["SQLAlchemy", "asyncpg"]
firestore = ["google-cloud-firestore "]

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest = "^6.1.2"
pytest-asyncio = "^0.14.0"
pytest-snapshot = "^0.8.0"
pytest-random-order = "^1.0.4"
async_asgi_testclient = "^1.4.4"
factory-boy = "^3.1.0"
invoke = "^1.5.0"
deepdiff = "^5.5.0"
alembic = "^1.4.3"
wait-for-it = "^2.2.1"
black = "^22.6.0"
mypy = "^0.971"
Faker = "^18.4.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
